refactor: split imports page into feature modules
